Documentation and License Added
This plugin is not published on Github, so we are publishing it in a
standard form, so that it can be discovered by other members of the
Movable Type community.

This commit adds documentation that explains this simple plugin more
completely.  It also adds a license so that people will know that they
are free to use it in any way that complies with the MIT License.

We credit Maarten Schenk for its initial development.